+++ project-base/app/translations/messages.cs.po
@@ -685,12 +685,6 @@ msgstr "Vyplňte také URL pro zvolenou doménu"
 msgid "Fill URL only for selected domain"
 msgstr "Vyplňte pouze URL pro zvolenou doménu"
 
-msgid "Filter parameters"
-msgstr "Parametry filtru"
-
-msgid "Filter parameters closed by default"
-msgstr "Defaultně zavřené parametry ve filtru"
-
 msgid "First name"
 msgstr "Jméno"
 
@@ -1161,9 +1155,6 @@ msgstr "Parametry - vše"
 
 msgid "Parameters - view"
 msgstr "Parametry - zobrazení"
-
-msgid "Parameters order in category"
-msgstr "Pořadí parametrů v kategorii"
 
 msgid "Pararameter value of type color"
 msgstr "Hodnota parametru typu barva"
