+++ project-base/app/translations/messages.en.po
@@ -685,12 +685,6 @@ msgstr ""
 msgid "Fill URL only for selected domain"
 msgstr ""
 
-msgid "Filter parameters"
-msgstr ""
-
-msgid "Filter parameters closed by default"
-msgstr ""
-
 msgid "First name"
 msgstr ""
 
@@ -1160,9 +1154,6 @@ msgid "Parameters - full"
 msgstr ""
 
 msgid "Parameters - view"
-msgstr ""
-
-msgid "Parameters order in category"
 msgstr ""
 
 msgid "Pararameter value of type color"
